backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
alembic==1.15.2
pydantic-settings==2.9.1
PyJWT[crypto]==2.10.1
httpx==0.28.1
